+++ enjoy/src/renderer/hooks/index.ts
@@ -1,4 +1,3 @@
-export * from './use-transcode';
+export * from './use-transcribe';
 export * from './use-ai-command';
 export * from './use-conversation';
-export * from './use-stt';
